PaymentContext/PaymentContext.Tests/Entities/StudentTests.cs
using Microsoft.VisualStudio.TestTools.UnitTesting;
using PaymentContext.Domain.Entities;
using PaymentContext.Domain.Enums;
using PaymentContext.Domain.ValueObjects;
using System;
using System.Collections.Generic;
using System.Text;

namespace PaymentContext.Tests.Entities
{
    [TestClass]
    public class StudentTests
    {
        private readonly Name _name;
        private readonly Document _document;
        private readonly Email _email;
        private readonly Address _address;

        private readonly Student _student;

        public StudentTests()
        {
            _name = new Name("Rodrigo", "Zanferrari");
            _document = new Document("32516143800", EDocumentType.CPF);
            _email = new Email("rodrigo.zanferrari@gmail.com");
            _address = new Address("Rua 1", "2", "Vila Augusta", "Guarulhos", "SP", "Brasil", "07023022");

            _student = new Student(_name, _document, _email);
        }

        [TestMethod]
        public void ShouldReturnErrorWhenHasActiveSubscription()
        {
            var subscription = new Subscription(null);
            var payment = new PayPalPayment("12345678", DateTime.Now, DateTime.Now.AddDays(5), 200, 200, "RZ2 Fotografia", _document, _email, _address);

            subscription.AddPayment(payment);
            _student.AddSubscription(subscription);
            _student.AddSubscription(subscription);

            Assert.IsTrue(!_student.IsValid);
        }

        [TestMethod]
        public void ShouldReturnSuccessWhenAddSubscription()
        {
            var subscription = new Subscription(null);
            var payment = new PayPalPayment("12345678", DateTime.Now, DateTime.Now.AddDays(5), 200, 200, "RZ2 Fotografia", _document, _email, _address);

            subscription.AddPayment(payment);
            _student.AddSubscription(subscription);

            Assert.IsTrue(_student.IsValid);
        }

        [TestMethod]
        public void ShouldReturnErrorWhenSubscriptionHasNoPayment()
        {
            var subscription = new Subscription(null);

            _student.AddSubscription(subscription);

            Assert.IsTrue(!_student.IsValid);
        }
    }
}
